+++ handler/src/query.rs
@@ -1,4 +1,4 @@
-use crate::handler::test_handler::TestQuery;
+use crate::test_handler::TestQuery;
 use async_graphql::MergedObject;
 
 #[derive(MergedObject, Default)]
