+++ server/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "server"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+handler = { path = "../handler" }
+context = { path = "../context" }
+async-graphql = "7.0.7"
+async-graphql-axum = "7.0.7"
+axum = "0.7.5"
+tokio = { version = "1.38.0", features = ["full"] }
