+++ server/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod server;
+
+pub(crate) mod schema;
