--- src/lib.rs
@@ -1,5 +1,0 @@
-mod context;
-mod handler;
-mod models;
-pub mod server;
-mod usecase;
