--- src/server.rs
@@ -1,2 +1,0 @@
-mod schema;
-pub mod server;
